requirements.txt
git+https://github.com/huggingface/transformers.git
sentencepiece
torch
torchaudio
datasets
soundfile
speechbrain
fastapi==0.95.0
pydantic==1.10.7
SQLAlchemy==2.0.9
python_bcrypt==0.3.2
